+++ packages/detect_proxy_setting/lib/detect_proxy_setting.dart
@@ -1,6 +1,5 @@
 import 'dart:async';
 
-import 'package:proxy_setting_platform_interface/proxy_setting.dart';
 import 'package:proxy_setting_platform_interface/proxy_setting_platform_interface.dart';
 
 Future<ProxySetting?> proxySetting({String? url}) async {