+++ packages/detect_proxy_setting/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+* add android
+
+## 0.0.2
+* add iOS
+
 ## 0.0.1
 * rename