+++ packages/proxy_setting_windows/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+* fix package setting
+
+## 0.0.2
+* fix package setting
+
 ## 0.0.1
 * rename